requirements.txt
# local package
-e .

# external requirements
python-git
pyqtgraph
PyOpenGL
PyOpenGL_accelerate
opensimplex
click
Sphinx
coverage
awscli
flake8
python-dotenv>=0.5.1


# backwards compatibility
pathlib2
